test(core): co-locate pipeline-frame suites by buildFrame concern
Split the 325-line pipeline-frame monolith into tests/pipeline-frame/
{identity,count,annotation,bin-advisory,remap-source-rows,stats,
regression,input-groups} matching already-modular pipeline/frame*.ts
production modules.

Fixes #553
